Rename the file that holds ToggleButton
This makes it more in keeping with other widgets, despite the fact that it
should be considered to be an internal widget.
