+++ src/textual/widgets/_toggle_button.py
